requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
jinja2
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]
bcrypt>=4.1.2
google-generativeai>=0.6.0
